pyproject.toml
[project]
name = "transformers-hw"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "foxglove-sdk>=0.14.3",
    "mcap>=1.3.0",
    "mcap-ros2-support>=0.5.5",
    "matplotlib>=3.7.0",
    "numpy>=2.2.6",
    "pin>=3.7.0",
    "pyarrow>=21.0.0",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "urdfpy>=0.0.22",
    "vamp-planner>=0.6.1",
]
